upload de mensagens pronto e policies encaminhadas

+++ app/Http/Controllers/BookController.php
@@ -6,7 +6,11 @@ use App\Models\Book;
 use App\Models\Author;
 use App\Models\Publisher;
 use App\Models\Category;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Validation\ValidationData;
+use Illuminate\Support\Facades\Storage;
 
 class BookController extends Controller
 {
@@ -27,6 +31,7 @@ class BookController extends Controller
     // Função para exibir o formulário de criação de um novo livro
     public function create()
     {
+        Gate::authorize('librarian', User::class);
         $authors = Author::all();
         $publishers = Publisher::all();
         $categories = Category::all();
@@ -36,14 +41,18 @@ class BookController extends Controller
     // Função para armazenar um novo livro no banco de dados
     public function store(Request $request)
     {
+        Gate::authorize('librarian', User::class);
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'author_id' => 'required|integer',
             'publisher_id' => 'required|integer',
             'published_year' => 'required|integer',
             'categories' => 'required|array',
+            'cover' => 'required|file|image'
         ]);
 
+        $bookCover = $validatedData['cover']->store('covers', 'public');
+        $validatedData['cover'] = $bookCover;
         $book = Book::create($validatedData);
         $book->categories()->attach($request->categories);
 
@@ -53,6 +62,7 @@ class BookController extends Controller
     // Função para exibir o formulário de edição de um livro
     public function edit($id)
     {
+        Gate::authorize('librarian', User::class);
         $book = Book::findOrFail($id);
         $authors = Author::all();
         $publishers = Publisher::all();
@@ -63,13 +73,18 @@ class BookController extends Controller
     // Função para atualizar um livro no banco de dados
     public function update(Request $request, $id)
     {
+        Gate::authorize('librarian', User::class);
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'author_id' => 'required|integer',
             'publisher_id' => 'required|integer',
             'published_year' => 'required|integer',
             'categories' => 'required|array',
+            'cover' => 'required|file|image'
         ]);
+
+        $bookCover = $validatedData['cover']->store('covers', 'public');
+        $validatedData['cover'] = $bookCover;
 
         $book = Book::findOrFail($id);
         $book->update($validatedData);
@@ -81,7 +96,9 @@ class BookController extends Controller
     // Função para excluir um livro do banco de dados
     public function destroy($id)
     {
+        Gate::authorize('librarian', User::class);
         $book = Book::findOrFail($id);
+        Storage::delete('public/' . $book->cover);
         $book->categories()->detach();
         $book->delete();
